feat [#7] : move AuthUserService.kt & Change Living UserRole
