+++ pullout-core/src/main/kotlin/com/pullout/app/services/AuthUserService.kt
@@ -1,4 +1,4 @@
-package com.pullout.app.services.impl
+package com.pullout.app.services
 
 import com.pullout.app.dto.AccountUserDto
