+++ pullout-core/src/main/kotlin/com/pullout/enums/LivingUserRole.kt
@@ -4,6 +4,6 @@ package com.pullout.enums
  * 건물 by 건물 사용자 룰
  */
 enum class LivingUserRole {
-    USER, // 입주 유저
-    MANAGER // 관리자
+    LIVING_USER, // 입주 유저
+    LIVING_MANAGER // 관리자
 }
